backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.62
langchain-google-genai==1.0.1
langchain-core==0.1.10
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
